Falcon api staging merge 6sep

+++ app/workers/inline_image_shredder.rb
@@ -8,12 +8,15 @@ class InlineImageShredder < BaseWorker
   def perform(args)
     args.symbolize_keys!
     path = DeletedBodyObserver.cleanup_file_path(Account.current.id, args[:model_name], args[:model_id])
+    return unless AwsWrapper::S3Object.exists?(path, S3_CONFIG[:bucket])
     content = AwsWrapper::S3Object.read(path, S3_CONFIG[:bucket])
     attachment_ids = get_attachment_ids(content)
     Account.current.attachments.where(id: attachment_ids).find_each(batch_size: 300) do |attachment|
       attachment.destroy if should_delete?(attachment.id, attachment.attachable_type)
     end
     AwsWrapper::S3Object.delete(path, S3_CONFIG[:bucket])
+  rescue AWS::S3::Errors::NoSuchKey => e
+    Rails.logger.error "Body Content not found in s3 :: #{Account.current.id} :: #{args.inspect}" 
   end
 
   private
